requirements.txt
Flask==2.3.2
Flask-OIDC==1.4.0
PyJWT==2.8.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
scipy==1.13.1
nolds==0.5.2
openpyxl==3.1.2
Werkzeug==3.0.2
